+++ test/conftest.py
@@ -98,6 +98,7 @@ def scigrid_de_network():
 def ac_dc_network_solved():
     n = pypsa.examples.ac_dc_meshed()
     n.optimize()
+    del n.model.solver_model
     return n